Codeclimate issue fix
