+++ app/src/main/java/org/ole/planet/myplanet/base/BaseContainerFragment.java
@@ -10,6 +10,7 @@ import android.support.v4.app.Fragment;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.AppCompatRatingBar;
 import android.util.Log;
+import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.MotionEvent;
 import android.view.View;
@@ -26,14 +27,21 @@ import com.google.gson.JsonObject;
 import org.ole.planet.myplanet.R;
 import org.ole.planet.myplanet.callback.OnHomeItemClickListener;
 import org.ole.planet.myplanet.callback.OnRatingChangeListener;
+import org.ole.planet.myplanet.model.RealmMeetup;
+import org.ole.planet.myplanet.model.RealmMyCourse;
 import org.ole.planet.myplanet.model.RealmMyLibrary;
+import org.ole.planet.myplanet.model.RealmMyLife;
+import org.ole.planet.myplanet.model.RealmMyTeam;
 import org.ole.planet.myplanet.service.UserProfileDbHandler;
 import org.ole.planet.myplanet.ui.calendar.CalendarFragment;
 import org.ole.planet.myplanet.ui.course.AdapterCourses;
+import org.ole.planet.myplanet.ui.course.TakeCourseFragment;
 import org.ole.planet.myplanet.ui.library.LibraryDetailFragment;
+import org.ole.planet.myplanet.ui.mymeetup.MyMeetupDetailFragment;
 import org.ole.planet.myplanet.ui.news.NewsFragment;
 import org.ole.planet.myplanet.ui.references.ReferenceFragment;
 import org.ole.planet.myplanet.ui.submission.MySubmissionFragment;
+import org.ole.planet.myplanet.ui.team.MyTeamsDetailFragment;
 import org.ole.planet.myplanet.ui.userprofile.AchievementFragment;
 import org.ole.planet.myplanet.ui.viewer.AudioPlayerActivity;
 import org.ole.planet.myplanet.ui.viewer.CSVViewerActivity;
@@ -49,6 +57,7 @@ import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 
+import io.realm.RealmObject;
 import io.realm.RealmResults;
 
 public abstract class BaseContainerFragment extends BaseResourceFragment {
@@ -237,38 +246,5 @@ public abstract class BaseContainerFragment extends BaseResourceFragment {
             });
         }
 
-    }
-
-    public void handleClick(final String id, String title, final Fragment f, TextView v) {
-        v.setText(title);
-        v.setOnClickListener(view -> {
-            if (homeItemClickListener != null) {
-                Bundle b = new Bundle();
-                b.putString("id", id);
-                f.setArguments(b);
-                homeItemClickListener.openCallFragment(f);
-            }
-        });
-    }
-
-    public void handleClickMyLife(final String id, String title, int imageId, LinearLayout linearLayout) {
-        ImageView imageView = (ImageView) linearLayout.getChildAt(0);
-        TextView textView = (TextView) linearLayout.getChildAt(1);
-        imageView.setImageResource(imageId);
-        textView.setText(title);
-        linearLayout.setOnClickListener(view -> {
-            if (homeItemClickListener != null) {
-                if (title.equals(getString(R.string.myhealth))) { Utilities.toast(getContext(), "Feature not available");
-                } else if (title.equals(getString(R.string.messeges))) { Utilities.toast(getContext(), "Feature not available");
-                } else if (title.equals(getString(R.string.submission))) { homeItemClickListener.openCallFragment(new MySubmissionFragment());
-                } else if (title.equals(getString(R.string.news))) { homeItemClickListener.openCallFragment(new NewsFragment());
-                } else if (title.equals(getString(R.string.references))) { homeItemClickListener.openCallFragment(new ReferenceFragment());
-                } else if (title.equals(getString(R.string.help_wanted))) { Utilities.toast(getContext(), "Feature not available");
-                } else if (title.equals(getString(R.string.calendar))) { homeItemClickListener.openCallFragment(new CalendarFragment());
-                } else if (title.equals(getString(R.string.contacts))) { Utilities.toast(getContext(), "Feature not available");
-                } else if (title.equals(getString(R.string.achievements))) { homeItemClickListener.openCallFragment(new AchievementFragment());
-                }
-            }
-        });
     }
 }
