+++ app/src/main/java/org/ole/planet/myplanet/ui/dashboard/BaseDashboardFragment.java
@@ -41,7 +41,7 @@ import io.realm.Case;
 import io.realm.Realm;
 import io.realm.RealmObject;
 
-public class BaseDashboardFragment extends BaseContainerFragment {
+public class BaseDashboardFragment extends BaseDashboardFragmentPlugin {
     String fullName;
     Realm mRealm;
     TextView txtFullName, txtVisits, txtRole;
@@ -125,69 +125,6 @@ public class BaseDashboardFragment extends BaseContainerFragment {
                 flexboxLayout.addView(myCoursesTextViewArray[itemCnt], params);
             }
             itemCnt++;
-        }
-    }
-
-    private void setTextColor(TextView textView, int itemCnt, Class c) {
-        //  int color = getResources().getColor(PreferenceManager.getDefaultSharedPreferences(getActivity()).getBoolean("bell_theme", false) ? Constants.COLOR_MAP.get(c) : R.color.md_grey_400);
-        textView.setTextColor(getResources().getColor(R.color.md_black_1000));
-        if ((itemCnt % 2) == 0) {
-            textView.setBackgroundResource(R.drawable.light_rect);
-        } else {
-            textView.setBackgroundColor(getResources().getColor(R.color.md_grey_300));
-        }
-    }
-
-    public void setLinearLayoutProperties(LinearLayout[] linearLayoutArray, int itemCnt, final RealmObject obj, Class c) {
-        linearLayoutArray[itemCnt] = new LinearLayout(getContext());
-        linearLayoutArray[itemCnt].setOrientation(LinearLayout.VERTICAL);
-        linearLayoutArray[itemCnt].setMinimumWidth(R.dimen.user_image_size);
-        linearLayoutArray[itemCnt].setGravity(Gravity.CENTER);
-        LinearLayout.LayoutParams lp = new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.MATCH_PARENT);
-        lp.gravity = Gravity.CENTER;
-        linearLayoutArray[itemCnt].setLayoutParams(lp);
-
-        TextView tv = new TextView(getContext());
-        tv.setGravity(Gravity.CENTER);
-        LinearLayout.LayoutParams lp_tv = new LinearLayout.LayoutParams(ViewGroup.LayoutParams.WRAP_CONTENT, ViewGroup.LayoutParams.WRAP_CONTENT);
-        lp_tv.gravity = Gravity.CENTER;
-        tv.setPadding(8, 8, 8, 8);
-
-        ImageView imageView = new ImageView(getContext());
-        LinearLayout.LayoutParams lp_iv = new LinearLayout.LayoutParams(36, 36);
-        lp_iv.gravity = Gravity.CENTER;
-        imageView.setLayoutParams(lp_iv);
-
-        linearLayoutArray[itemCnt].addView(imageView);
-        linearLayoutArray[itemCnt].addView(tv);
-
-        tv.setTextColor(getResources().getColor(R.color.md_black_1000));
-        if ((itemCnt % 2) == 0) {
-            linearLayoutArray[itemCnt].setBackgroundResource(R.drawable.light_rect);
-        } else {
-            linearLayoutArray[itemCnt].setBackgroundColor(getResources().getColor(R.color.md_grey_300));
-        }
-
-        handleClickMyLife(((RealmMyLife) obj).get_id(), ((RealmMyLife) obj).getTitle(), ((RealmMyLife) obj).getImageId(), linearLayoutArray[itemCnt]);
-
-    }
-
-    public void setTextViewProperties(TextView[] textViewArray, int itemCnt, final RealmObject obj, Class c) {
-        textViewArray[itemCnt] = new TextView(getContext());
-        textViewArray[itemCnt].setPadding(20, 10, 20, 10);
-        textViewArray[itemCnt].setTextAlignment(View.TEXT_ALIGNMENT_CENTER);
-        textViewArray[itemCnt].setGravity(Gravity.CENTER_VERTICAL | Gravity.CENTER_HORIZONTAL);
-        if (obj instanceof RealmMyLibrary) {
-            textViewArray[itemCnt].setText(((RealmMyLibrary) obj).getTitle());
-        } else if (obj instanceof RealmMyCourse) {
-            handleClick(((RealmMyCourse) obj).getCourseId(), ((RealmMyCourse) obj).getCourseTitle(), new TakeCourseFragment(), textViewArray[itemCnt]);
-        } else if (obj instanceof RealmMyTeam) {
-            //    textViewArray[itemCnt].setText(((RealmMyTeam) obj).getName());
-            handleClick(((RealmMyTeam) obj).getId(), ((RealmMyTeam) obj).getName(), new MyTeamsDetailFragment(), textViewArray[itemCnt]);
-        } else if (obj instanceof RealmMeetup) {
-            handleClick(((RealmMeetup) obj).getMeetupId(), ((RealmMeetup) obj).getTitle(), new MyMeetupDetailFragment(), textViewArray[itemCnt]);
-        } else if (obj instanceof RealmMyLife) {
-            // handleClickMyLife(((RealmMyLife) obj).get_id(),((RealmMyLife) obj).getTitle(),((RealmMyLife) obj).getImageId(),li[itemCnt]);
         }
     }
